+++ kotlin-compiler-extensions/test/kotlin/com/javiersc/kotlin/compiler/extensions/FakeIrExtension.kt
@@ -1,10 +1,10 @@
 package com.javiersc.kotlin.compiler.extensions
 
 import com.javiersc.kotlin.compiler.extensions.ir.IrTreeNode
-import com.javiersc.kotlin.compiler.extensions.ir.buildIrReturn
 import com.javiersc.kotlin.compiler.extensions.ir.name
 import com.javiersc.kotlin.compiler.extensions.ir.treeNode
 import com.javiersc.kotlin.compiler.extensions.ir.asIr
+import com.javiersc.kotlin.compiler.extensions.ir.copyIrReturn
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.ir.declarations.IrModuleFragment
@@ -27,7 +27,7 @@ class FakeIrExtension : IrGenerationExtension {
                 .asIr<IrCall>()
         val greetingsIrSimpleFunction: IrSimpleFunction = greetingsIrCall?.symbol?.owner ?: return
         val hiIrExpression: IrConst<*> = "hi".toIrConst(builtIns.stringType)
-        val hiIrReturn: IrReturn = greetingsIrSimpleFunction.buildIrReturn(hiIrExpression)
+        val hiIrReturn: IrReturn = greetingsIrSimpleFunction.copyIrReturn(hiIrExpression)
         greetingsIrSimpleFunction.body.asIr<IrBlockBody>()?.statements?.apply {
             clear()
             add(hiIrReturn)